MOVE 'set_database_config' to persistence module
