+++ app/websocket/app/main/persistence/utils.py
@@ -1,16 +1,55 @@
 from .models import Worker as WorkerMDL
 from .models import WorkerObject, TorchModel
 from .models import db
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
 
 import torch
 from syft.messaging.plan import Plan
 import syft as sy
-
+import os
 
 # Cache keys already saved in database.
 # Used to make snapshot more efficient.
 last_snapshot_keys = set()
 
+def set_database_config(app, test_config=None, verbose=False):
+    """ Set configs to use SQL Alchemy library.
+
+        Args:
+            app: Flask application.
+            test_config : Dictionary containing SQLAlchemy configs for test purposes.
+            verbose : Level of flask application verbosity.
+        Returns:
+            app: Flask application.
+        Raises:
+            RuntimeError : If DATABASE_URL or test_config didn't initialized, RuntimeError exception will be raised.
+    """
+    global db
+    db_url = os.environ.get("DATABASE_URL")
+    migrate = Migrate(app, db)
+    if test_config is None:
+        if db_url:
+            app.config.from_mapping(
+                SQLALCHEMY_DATABASE_URI=db_url, SQLALCHEMY_TRACK_MODIFICATIONS=False
+            )
+        else:
+            raise RuntimeError(
+                "Invalid database address : Set DATABASE_URL environment var or add test_config parameter at create_app method."
+            )
+    else:
+        app.config["SQLALCHEMY_DATABASE_URI"] = test_config["SQLALCHEMY_DATABASE_URI"]
+        app.config["TESTING"] = (
+            test_config["TESTING"] if test_config.get("TESTING") else True
+        )
+        app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = (
+            test_config["SQLALCHEMY_TRACK_MODIFICATIONS"]
+            if test_config.get("SQLALCHEMY_TRACK_MODIFICATIONS")
+            else False
+        )
+    app.config["VERBOSE"] = verbose
+    db.init_app(app)
+    return app
 
 def snapshot(worker):
     """ Take a snapshot of worker's current state.
